% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{cov.matrix}
\alias{cov.matrix}
\title{Cov helper function}
\usage{
\method{cov}{matrix}(x, ..., covEst = stats::cov)
}
\arguments{
\item{x}{data}

\item{...}{other options passed to method}

\item{method}{method of cov/precision estimation}
}
\description{
Cov helper function
}
\examples{
cov(iris[,1:4])
}
\keyword{internal}
