% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreesFreedom.R
\name{degreesFreedom}
\alias{degreesFreedom}
\title{Degrees of Freedom of a Covariance Matrix}
\usage{
degreesFreedom(x)
}
\arguments{
\item{x}{a matrix with class \code{covariance}}
}
\value{
The degrees of freedom of the covariance matrix, as a \code{numeric}
   object
}
\description{
Degrees of Freedom of a Covariance Matrix
}
\details{
This function uses the \code{f_eval} function from the \code{lazyeval}
   package to return the degrees of freedom stored within the \code{df}
   attribute of a \code{matrix} object having the additional class
   \code{covariance}. This evaluation is necessary because the \code{df}
   attribute of a matrix with class \code{covariance} is stored as a
   \code{formula} object.
}
\examples{
degreesFreedom(cov(iris[,1:4]))
lapply(cov(iris, group = "Species"), degreesFreedom)
}
