% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_cov.R
\name{pooled_cov.matrix}
\alias{pooled_cov.matrix}
\title{Cov helper function}
\usage{
\method{pooled_cov}{matrix}(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{other options passed to method}

\item{method}{method of cov/precision estimation}
}
\description{
Cov helper function
}
\examples{
cov(iris[,1:4])
}
\keyword{internal}
