% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_cov.R
\name{pooled_cov}
\alias{pooled_cov}
\title{Pooled Covariance}
\usage{
pooled_cov(x, ..., covEst = stats::cov)
}
\arguments{
\item{x}{data}

\item{...}{other options passed to estimation method}

\item{covEst}{covariance matrix estimation method}
}
\value{
pooled covariance class matrix
}
\description{
Pooled Covariance
}
\examples{
pooled_cov(iris, group = Species)
}
