% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{cov}
\alias{cov}
\title{Covariance Matrix}
\usage{
cov(x, ..., covEst = stats::cov)
}
\arguments{
\item{x}{data}

\item{...}{other options passed to covarince estimation method}

\item{covEst}{covariance estimation method}
}
\value{
list of covariance class matries
}
\description{
Covariance Matrix
}
\examples{
cov(iris[,1:4])
}
