% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreesFreedom.R
\name{degreesFreedom}
\alias{degreesFreedom}
\title{Title}
\usage{
degreesFreedom(x)
}
\arguments{
\item{x}{covariance class matrix}
}
\value{
Degrees of freedom of the covariance matrix
}
\description{
Title
}
\examples{
degreesFreedom(cov(iris[,1:4]))
}
