% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_cov.R
\name{pooled_cov}
\alias{pooled_cov}
\title{Pooled Covariance}
\usage{
pooled_cov(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{other options}
}
\description{
Pooled Covariance
}
\examples{
pooled_cov(iris, group = Species)
}
